requirements.txt
# Framework to scrape OPGG website
opgg.py
# MySQL DB connector with python
mysql-connector-python
# Pandas
pandas
# type checking
mypy